+++ app/thirdparty/java/other/writeily/widget/WrWidgetConfigure.java
@@ -19,7 +19,7 @@ import android.support.v4.app.FragmentManager;
 import android.support.v7.app.AppCompatActivity;
 
 import net.gsantner.markor.R;
-import net.gsantner.markor.ui.FilesystemViewerFactory;
+import net.gsantner.markor.ui.FilesystemViewerCreator;
 import net.gsantner.markor.util.AppSettings;
 import net.gsantner.markor.util.PermissionChecker;
 import net.gsantner.opoc.ui.FilesystemViewerData;
@@ -44,7 +44,7 @@ public class WrWidgetConfigure extends AppCompatActivity {
         PermissionChecker permc = new PermissionChecker(this);
         if (permc.mkdirIfStoragePermissionGranted()) {
             FragmentManager fragManager = getSupportFragmentManager();
-            FilesystemViewerFactory.showFolderDialog(new FilesystemViewerData.SelectionListenerAdapter() {
+            FilesystemViewerCreator.showFolderDialog(new FilesystemViewerData.SelectionListenerAdapter() {
                 @Override
                 public void onFsViewerSelected(String request, File file) {
                     complete(file.getAbsolutePath());